fix: do not use a distinct color for visited
Many well-known sites do not use a distinct color for visited anymore.
The main problem for fx with it is that purple doesn't look nice on a
dark background. Also, I think in the context of blogging it is probably
noise. The information is in the text and brightly colored visit
information is a distraction.