fix: check required attribute before disabling submit button (#89)
When setting up a new blog in production, the textareas are blank in the
`Settings` page.

This disables the `Save` button because of this logic: 
https://github.com/rikhuijzer/fx/blob/main/fx/src/static/nodefer.js#L39

Moreoever, those textareas don't have an `oninput` hander like we do
here:
https://github.com/rikhuijzer/fx/blob/b71dc5c71fd0478839aa5f6ee7738de83161e988/fx/src/html.rs#L308

Since the textareas on the `Settings` page are optional, I didn't want
to add this `oninput` handler to them.

Instead, I've modified the disabling logic to disable the submit button
if a _required_ textarea is empty. And I've added the required attribute
to the `post` textarea, which was needed anyhow.